requirements.txt
RPi.GPIO==0.7.0
python-socketio[client]
statistics
simpleaudio
pydub
